RAT_client/requirements.txt
colorama>=0.4.6
Pillow>=10.0.0